* commit '3fbfc65abfeb5245d5d6e47e628a535ff03a8abf': (52 commits)
  fix all tests
  fix boardexport
  refactoring done
  refac unfinished
  fix testsss
  fix html 404 error
  fix export board test
  fix publicVsPrivateBoards
  fix testsss
  correct test
  add test
  fix all tests
  x
  fix some tests
  fix transfer steward test
  fix classify all form in formboard
  fix tests and fixing btn in form board
  fix classify all cdes in board
  fix some tests
  fix registrationCtrl
  ...

+++ modules/system/node-js/app.js
@@ -1,5 +1,6 @@
 var passport = require('passport')
     , mongo_data_system = require('./mongo-data')
+    , mongo_cde = require('../../cde/node-js/mongo-cde')
     , config = require('./parseConfig')
     , dbLogger = require('./dbLogger.js')
     , logging = require('./logging.js')
@@ -10,7 +11,6 @@ var passport = require('passport')
     , classificationShared = require('../shared/classificationShared.js')
     , classificationNode = require('./classificationNode')
     , adminItemSvc = require("./adminItemSvc")
-    , auth = require('./authorization')
     , csrf = require('csurf')
     , authorizationShared = require("../../system/shared/authorizationShared")
     , daoManager = require('./moduleDaoManager')
@@ -20,7 +20,7 @@ var passport = require('passport')
     , tar = require('tar-fs')
     , zlib = require('zlib')
     , spawn = require('child_process').spawn
-    , authorization = require('../../system/node-js/authorization')
+    , authorization = require('./authorization')
     , esInit = require('./elasticSearchInit')
     , elastic = require('./elastic.js')
     , app_status = require("./status.js")
@@ -46,7 +46,7 @@ exports.init = function (app) {
 
     ["/cde/search", "/form/search", "/home", "/stats", "/help/:title", "/createForm", "/createCde", "/boardList",
         "/board/:id", "/deview", "/myboards", "/sdcview",
-        "/cdeStatusReport", "/board/:id", "/deview", "/myboards", "/sdcview",
+        "/cdeStatusReport", "/deview", "/myboards", "/sdcview",
         "/formView", "/quickBoard", "/searchSettings", "/siteAudit", "/siteaccountmanagement", "/orgaccountmanagement",
         "/classificationmanagement", "/inbox", "/profile", "/login", "/orgAuthority"].forEach(function (path) {
         app.get(path, function (req, res) {
@@ -218,30 +218,30 @@ exports.init = function (app) {
                 function checkCaptcha(captchaDone) {
                     // disabled for now.
                     return captchaDone();
-                    if (failedIp && failedIp.nb > 2) {
-                        if (req.body.recaptcha) {
-                            request.post("https://www.google.com/recaptcha/api/siteverify",
-                                {
-                                    form: {
-                                        secret: config.captchaCode,
-                                        response: req.body.recaptcha,
-                                        remoteip: getRealIp(req)
-                                    },
-                                    json: true
-                                }, function (err, resp, body) {
-                                    if (err) captchaDone(err);
-                                    else if (!body.success) {
-                                        captchaDone("incorrect recaptcha");
-                                    } else {
-                                        captchaDone();
-                                    }
-                                });
-                        } else {
-                            captchaDone("missing reCaptcha");
-                        }
-                    } else {
-                        captchaDone();
-                    }
+                    //if (failedIp && failedIp.nb > 2) {
+                    //    if (req.body.recaptcha) {
+                    //        request.post("https://www.google.com/recaptcha/api/siteverify",
+                    //            {
+                    //                form: {
+                    //                    secret: config.captchaCode,
+                    //                    response: req.body.recaptcha,
+                    //                    remoteip: getRealIp(req)
+                    //                },
+                    //                json: true
+                    //            }, function (err, resp, body) {
+                    //                if (err) captchaDone(err);
+                    //                else if (!body.success) {
+                    //                    captchaDone("incorrect recaptcha");
+                    //                } else {
+                    //                    captchaDone();
+                    //                }
+                    //            });
+                    //    } else {
+                    //        captchaDone("missing reCaptcha");
+                    //    }
+                    //} else {
+                    //    captchaDone();
+                    //}
                 }],
             function allDone(err) {
                 if (err) {
@@ -538,58 +538,7 @@ exports.init = function (app) {
             res.send(status);
         });
     });
-
-    app.post('/classification/elt', function (req, res) {
-        if (!usersrvc.isCuratorOf(req.user, req.body.orgName)) {
-            res.status(401).send();
-            return;
-        }
-        classificationNode.cdeClassification(req.body, classificationShared.actions.create, function (err) {
-            if (!err) {
-                res.send({code: 200, msg: "Classification Added"});
-                mongo_data_system.addToClassifAudit({
-                    date: new Date()
-                    , user: {
-                        username: req.user.username
-                    }
-                    , elements: [{
-                        _id: req.body.cdeId
-                    }]
-                    , action: "add"
-                    , path: [req.body.orgName].concat(req.body.categories)
-                });
-            } else {
-                res.send({code: 403, msg: "Classification Already Exists"});
-            }
-
-        });
-    });
-
-    app.delete('/classification/elt', function (req, res) {
-        if (!usersrvc.isCuratorOf(req.user, req.query.orgName)) {
-            res.status(401).send();
-            return;
-        }
-        classificationNode.cdeClassification(req.query, classificationShared.actions.delete, function (err) {
-            if (!err) {
-                res.end();
-                mongo_data_system.addToClassifAudit({
-                    date: new Date()
-                    , user: {
-                        username: req.user.username
-                    }
-                    , elements: [{
-                        _id: req.query.cdeId
-                    }]
-                    , action: "delete"
-                    , path: [req.query.orgName].concat(req.query.categories)
-                });
-            } else {
-                res.status(202).send({error: {message: "Classification does not exists."}});
-            }
-        });
-    });
-
+    
     app.delete('/classification/org', function (req, res) {
         if (!usersrvc.isCuratorOf(req.user, req.query.orgName)) {
             res.status(403).end();
@@ -638,6 +587,7 @@ exports.init = function (app) {
         orgsvc.transferSteward(req, res);
     });
 
+    // TODO this works only for CDEs. Forms TODO later.
     app.post('/classification/bulk/tinyid', function (req, res) {
         if (!usersrvc.isCuratorOf(req.user, req.body.classification.orgName)) {
             res.status(403).send("Not Authorized");
@@ -650,7 +600,7 @@ exports.init = function (app) {
                 , tinyId: elt.id || elt
                 , version: elt.version || null
             };
-            classificationNode.cdeClassification(classifReq, classificationShared.actions.create, actionCallback);
+            classificationNode.eltClassification(classifReq, classificationShared.actions.create, mongo_cde, actionCallback);
         };
         adminItemSvc.bulkAction(req.body.elements, action, function () {
             var elts = req.body.elements.map(function (e) {
@@ -677,7 +627,7 @@ exports.init = function (app) {
     });
 
     app.get('/getAllUsernames', function (req, res) {
-        if (auth.isSiteOrgAdmin(req)) {
+        if (authorization.isSiteOrgAdmin(req)) {
             usersrvc.getAllUsernames(req, res);
         } else {
             res.status(401).send();
@@ -802,7 +752,8 @@ exports.init = function (app) {
     app.get('/attachment/decline/:id', function (req, res) {
         if (!authorizationShared.hasRole(req.user, "AttachmentReviewer")) return res.status(401).send();
         daoManager.getDaoList().forEach(function (dao) {
-            dao.removeAttachmentLinks(req.params.id);
+            if (dao.removeAttachmentLinks)
+                dao.removeAttachmentLinks(req.params.id);
         });
         mongo_data_system.deleteFileById(req.params.id);
         res.send("Attachment declined");
@@ -1062,8 +1013,8 @@ exports.init = function (app) {
     }).start();
 
 
-    app.get('/comments/tinyId/:tinyId', function (req, res) {
-        mongo_data_system.Comment.find({"element.eltId": req.params.tinyId}).sort({created: 1}).exec(function(err, comments) {
+    app.get('/comments/eltId/:eltId', function (req, res) {
+        mongo_data_system.Comment.find({"element.eltId": req.params.eltId}).sort({created: 1}).exec(function(err, comments) {
             var result = comments.filter(c => c.status !== 'deleted');
             result.forEach(function (c) {
                 c.replies = c.replies.filter(r => r.status !== 'deleted');
